29 player screen enhancements

+++ app/src/main/java/com/imdmp/youtubecompose/features/videoplayer/comments/CommentState.kt
@@ -12,6 +12,6 @@ data class CommentModel(
     val authorName: String,
     val commentText: String,
     val profilePic: String,
-    val likeCount: Long,
+    val likeCount: Int,
     val timeCommented: String,
 )
